% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_states_response}
\alias{parse_states_response}
\title{Parse OpenSky state vectors response}
\usage{
parse_states_response(x)
}
\arguments{
\item{x}{Response from OpenSky states endpoint}
}
\value{
A tibble of state vectors
}
\description{
Parse OpenSky state vectors response
}
\keyword{internal}
